Libft

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mmkhwana <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/30 08:32:25 by mmkhwana          #+#    #+#             */
-/*   Updated: 2019/06/30 13:05:01 by mmkhwana         ###   ########.fr       */
+/*   Updated: 2019/06/30 16:11:47 by mmkhwana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ void				*ft_memcpy(void *dst, const void *str, size_t n);
 void				*ft_memmove(void *dst, const void *str, size_t n);
 char				*ft_strchr(const char *str, int c);
 char				*ft_strcpy(char *dst, const char *str);
-char				*ft_strncat(char *str1, const char *str2, int n);
+char				*ft_strncat(char *str1, const char *str2, size_t n);
 char				*ft_strncpy(char *strd, const char *strs, size_t n);
 int					ft_memcmp(const void *str1, const void *str2, size_t n);
 void				*ft_memchr(const void *str1, int c, size_t n);
